pyproject.toml
[tool.poetry]
name = "ai-ledger"
version = "0.1.0"
description = "AI Ledger - financial tracker"
authors = [
    "max1m111an <maksimilianbegunov@inbox.ru>",
    "PewPewP <na.aleksandrova2014@gmail.com>",
    "Qurlya <pingvining05@mail.ru>"
]
readme = "README.md"
packages = [{include = "apps"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
bcrypt = "^4.3.0"
pydantic = "^2.11.9"
starlette = "^0.48.0"
fastapi-another-jwt-auth = "^0.1.11"
cffi = "^2.0.0"
sqlmodel = "^0.0.25"
aiomysql = "^0.2.0"
passlib = "^1.7.4"
platformdirs = "^4.2.0"
virtualenv = "^20.35.3"
pre-commit = "^4.3.0"
python-multipart = "^0.0.21"
opencv-python = "^4.12.0.88"
pillow = "^12.0.0"
pytesseract = "^0.3.13"
paddleocr = "^3.3.2"
paddlepaddle = "^3.2.2"
imutils = "^0.5.4"
rapidfuzz = "^3.14.3"
regex = "^2025.11.3"
rich = "^14.2.0"
ollama = "^0.6.1"
numpy = ">=2,<2.3"
scipy = "^1.16.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
